Engine: BlendMode fixes

+++ Engine/ac/characterextras.h
@@ -44,7 +44,7 @@ struct CharacterExtras {
     char  process_idle_this_time;
     char  slow_move_counter;
     short animwait;
-    short blend_mode;
+    int blend_mode;
 
     void ReadFromFile(Common::Stream *in, int32_t cmp_ver);
     void WriteToFile(Common::Stream *out);
